move all form logic into form component
